+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### Added
+ - waspdrive scheduler capability.
+
 ## [1.0.3] - 10-07-2019
 ### Added
 - GCC 5.4 compiler compatibility CI job.
@@ -11,7 +14,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - waspdakota: a package for the Dakota UQ coupling. Specifically, support for the Dakota Parameter file format
 - wasputils/substitute: a utility that allows direct use of the Dakota parameter file for template attribute substitution
 - HIVE UNKNOWN input node validation capability (See HIVE readme for details).
-
 
 ### Changed
 - wasvii hanging comment (comment 1 or more lines below last parameter) logic to reparent to the block